Fixed the progress bar. If you don't convert to float before doing the division, it will do an integer division always resulting in false.
git-svn-id: svn://svn.code.sf.net/p/themis/code@587 ecaf47ba-bb13-0410-a3b6-899b003a5108

+++ trunk/themis/framework/SiteHandler.cpp
@@ -313,7 +313,7 @@ status_t SiteHandler :: ReceiveBroadcast(BMessage* msg) {
 						int8 loadprogress = -1;
 					
 						if (content_length != 0) {
-							loadprogress = (int8)((bytes_received / content_length) * 100);
+							loadprogress = (int8)(((float)bytes_received / (float)content_length) * 100);
 						}
 						else { // chunked transfer mode
 							loadprogress = 50;
